Adding test and renaming file
